+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Unreleased
 ----------
-- Added `Debug` impl for `CacheClient`
+- Added `CacheClient::from_env` constructor
+- Added `Debug` impl for `CachingClient`
 
 
 0.1.0